Tell a registry outage from a finding in the CI audit gate too
This pull request went red on a green tree: 373 suites and 7,437 tests
passed, lint passed, coverage passed, and the job failed because
`npm audit` got a 503 from the registry's bulk advisory endpoint.

`npm audit` exits non-zero for two unrelated reasons — it found
advisories at or above the level, or it never managed to ask — and
ci.yml's step read the two alike, so an npmjs blip arrives as a security
finding with nothing in the log to say the tree simply was not audited.
That is the conflation #962 fixed on security-scan.yml's weekly job. The
same script is now on the pull-request gate, where it matters more: this
is the gate everyone has to merge past, and a gate that cries wolf is one
people learn to merge past.

Registry failures — npm's own "audit endpoint returned an error", plus
the network codes — are retried three times with backoff. Everything else
is a real report and fails immediately, unretried. If the registry never
answers the job still fails, and says which it was: an un-audited tree is
not a clean one.

The two steps are deliberately one script. They already had to stay at
the same bar, and the cases that hold them there are now shared:
tests/helpers/auditStep.js runs a step for real against a stubbed npm —
matching on npm's words rather than the HTTP code, since "503" and
"Service Unavailable" both turn up inside advisory titles and version
ranges — and both workflows' steps are driven through the same six cases,
so a copy that drifts is a copy that fails. The weekly job's own tests
now come from there rather than from a private copy.
